modularization for shared files (RAG)
